+++ Sorting/selection_sort.cpp
@@ -4,13 +4,13 @@ using std::cout, std::endl;
 
 // COMPLEXITY: Time - O(n^2)
 
-void selectionSort(int array[], size_t length) {
+void selectionSort(int array[], int length) {
 	int minIndex = 0, tempValue = 0;
 
-	for (size_t i = 0; i < length; i++) {
+	for (int i = 0; i < length; i++) {
 		minIndex = i;
 		// iterates to obtain minimum index
-		for (size_t j = i + 1; j < length; j++) {
+		for (int j = i + 1; j < length; j++) {
 			if (array[j] < array[minIndex]) {
 				minIndex = j;
 			}
@@ -29,15 +29,15 @@ void selectionSort(int array[], size_t length) {
 
 
 
-void printArray(int array[], size_t length) {
-	for (size_t i = 0; i < length - 1; i++) {
+void printArray(int array[], int length) {
+	for (int i = 0; i < length - 1; i++) {
 		cout << array[i] << " -- ";
 	}
 	cout << array[length - 1] << endl;
 }
 
 int main() {
-	const size_t arrayLength = 10;
+	const int arrayLength = 10;
 	int array[arrayLength] = {1, 5, 3, 8, 6, 7, 9, 2, 10, 4};
 
 	cout << "INTIAL ARRAY:" << endl;
